requirements.txt
discord.py>=2.6.0,<3.0
python-dotenv>=1.0.0,<2.0.0
httpx>=0.27.0
openai>=1.50.0